pyproject.toml
[project]
name = "onepw-receiver"
version = "0.1.0"
description = ""
authors = ["fant5y <vroni@salala.de>"]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "onepasswordconnectsdk ~= 1",
    "rich ~= 13",
    "python-dotenv",
    "tomlkit ~= 0.12.4"
]

[project.optional-dependencies]
black = { extras = ["d"], version = "^24" }
isort = "^5"
mypy = "^1.9"
pipreqs = "^0.5"
pylint = "^3.1.0"


[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88
extra_standard_library = ["pytest"]

[mypy]
warn_return_any = false
warn_unused_configs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_column_numbers = true
show_error_codes = true
show_error_context = true

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"
